requirements.txt
#diffusers==0.15.0  已有
jieba==0.42.1
# matplotlib==3.7.1  已有。相同版本
# numpy==1.22.3  已有 1.22.2
# opencv_python==4.7.0.72  已有'4.6.0'
# pandas==2.0.0  已有1.5.1
Pillow==9.5.0
#psutil==5.9.0  已有'5.9.4'
# pydantic==1.10.7 已有，版本同
# scipy==1.10.1 已有，版本同
#setuptools==65.5.0 # 已有。'65.5.1'
timm==0.4.12
# torch==1.13.1     已有。目前是'2.1.0a0+fe05266'
torchscale==0.2.0
#torchvision==0.14.1      # 目前是'0.15.0a0'
# tqdm==4.64.1            # 目前是'4.65.0'
#transformers==4.28.0     # 目前是'4.31.0'
typing_extensions==4.5.0
bminf
